完成周期设置 功能模块

+++ Application/Cycle/Controller/IndexController.class.php
@@ -2,6 +2,7 @@
 namespace Cycle\Controller;
 use Admin\Controller\AdminController;
 use Cycle\Logic\CycleLogic;				//周期
+use Project\Logic\ProjectLogic;		//项目
 class IndexController extends AdminController
 {
 	public function indexAction()
@@ -24,14 +25,19 @@ class IndexController extends AdminController
 			//初始化
 			$id = (int)I('get.id');
 			$cycle = null;
-
+			
+			$CycleL = new CycleLogic();
 			//存在ID，即为编辑，取出当前周期数据
 			if($id)
 			{
-				$CycleL = new CycleLogic();
-				$cycle = getListById($id);
+				$cycle = $CycleL->getListById($id);
 			}
 
+			//取出所有父周期
+			$rootLists = $CycleL->getRootLists();
+
+			$this->assign("rootLists",$rootLists);
+			$this->assign("cycle",$cycle);
 			$this->assign('YZBODY',$this->fetch());
 			$this->display(YZTemplate);	
 		}
@@ -49,6 +55,7 @@ class IndexController extends AdminController
 		try
 		{
 			$CycleL = new CycleLogic();
+			dump(I('post.'));
 			//存在ID，则说明是更新操作
 			if($id)
 			{
@@ -79,7 +86,55 @@ class IndexController extends AdminController
 		$CycleL = new CycleLogic();
 		try
 		{
+			//再看自己否是有子周期。
+			if($CycleL->getListByPid($id))
+			{
+				$this->error("该周期下挂有子周期，请先删除子周期",U("index?p=" . I('get.p')));
+				return;
+			}
+
+			//先检查是否有项目已经添加到了这个周期上。
+			$ProjectL = new ProjectLogic();
+			$project = $ProjectL->getListsByCycleId($id);
+			if(!empty($project))
+			{
+				$this->error("已经有项目添加到该周期上，请先修改或删除相关项目",U("index?p=" . I('get.p')));				
+				return;
+			}
+
 			$CycleL->deleteById($id);
+			$this->success("操作成功",U('index?p='.I('get.p')));
+		}
+		catch(\Think\Exception $e)
+		{
+			$this->error = $e;
+			$this->_empty();
+		}
+
+	}
+
+	/**
+	 * 设某一周期 设置为当前统计周期
+	 */
+	public function setCurrentAction()
+	{
+		$id = (int)I('get.id');
+		try
+		{
+			//取出当前记录，没有报错。
+			$CycleL = new CycleLogic();
+			if( !$cycle = $CycleL->getListById($id))
+			{
+				E("传入的ID值有误，未找到相关记录");
+			}
+
+			//删除原有的当前周期,实为更新操作
+			$CycleL->deleteCurrentLists();
+
+			//添加本条周期 
+			$CycleL->setCurrentListById($id);
+
+			$this->success("操作成功",U('index?p=' . I('get.p')));
 		}
 		catch(\Think\Exception $e)
 		{
